cantondex-backend/api-gateway/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiohttp==3.9.1
asyncio==3.4.3
python-dateutil==2.8.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
web3==6.15.0
eth-account==0.10.0
PyJWT==2.8.0
